Move load user to crud
